requirements.txt
# Core Framework
fastapi==0.110.0
uvicorn[standard]==0.40.0
pydantic==2.11.9

# Database & ORM (optional, for future features)
sqlalchemy==2.0.23

# AI/LLM
openai==1.83.0
crewai[google-genai]==1.8.0

# AWS
boto3==1.34.1
botocore==1.34.1

# Environment
python-dotenv==1.1.1
python-multipart==0.0.21

# Monitoring & Logging
python-json-logger==2.0.7

# Utilities
requests==2.31.0
aiohttp==3.9.1

# Development
pytest==7.4.3
black==23.12.0
flake8==6.1.0
